\name{bock}
\alias{bock}
\docType{data}
\encoding{latin1}
\title{ Gene Expression Matrix of Published Data }
\description{
  This is a matrix object containing genes as rows and the samples as columns, where each element
  in the matrix is an expression value. The dataset contains 12900 genes and 32 samples.
}
\usage{data(bock)}
\value{
  A matrix with expression values for 12900 genes probes on 32 samples.
}
\seealso{
\code{\link{pways.kegg}}, \code{\link{pways.reactome}}
}
\examples{
data(bock)
}
\keyword{datasets}