\name{pways.kegg}
\alias{pways.kegg}
\docType{data}
\encoding{latin1}
\title{List containing pathway IDs, names, and genes in each pathway}
\description{
  This is a list containing a \code{vector} of pathway IDs, a \code{vector} of pathway names, and a list of the genes in each pathway.
  There are a total of 272 KEGG pathways.
}
\usage{data(pways.kegg)}
\value{
  A list containing a \code{vector} of pathway IDs, a \code{vector} of pathway names, and a list of the genes in each pathway for 272 pathways.
}
\seealso{
\code{\link{bock}}, \code{\link{pways.reactome}}
}
\examples{
data(pways.kegg)
}
\keyword{datasets}