\name{pways.reactome}
\alias{pways.reactome}
\docType{data}
\encoding{latin1}
\title{List containing pathway IDs, names, and genes in each pathway}
\description{
  This is a list containing a \code{vector} of pathway IDs, a \code{vector} of pathway names, and a list of the genes in each pathway.
  There are a total of 946 reactome pathways.
}
\usage{data(pways.reactome)}
\value{
  A list containing a \code{vector} of pathway IDs, a \code{vector} of pathway names, and a list of the genes in each pathway for 946 pathways.
}
\seealso{
\code{\link{bock}}, \code{\link{pways.kegg}}
}
\examples{
data(pways.reactome)
}
\keyword{datasets}